requirements.txt
fastapi
uvicorn
python-multipart
pillow
pytesseract
pydantic
pydantic-settings
python-dotenv
aiofiles